1021 실습코드 2

+++ day03/src/main/java/edu/sm/app/dto/ItemDto.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.time.LocalDate;
 import java.time.LocalDateTime;
 
 @Data
@@ -12,9 +13,9 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @NoArgsConstructor
 public class ItemDto {
-    private int  itemId;
+    private int itemId;
     private String itemName;
-    private long itemPrice;
+    private int itemPrice;
     private String imgName;
     private LocalDateTime regDate;
     private LocalDateTime updateDate;
